extract rocketlogger data and log file handling
