+++ software/node_server/app.js
@@ -36,7 +36,6 @@ import fs from 'fs/promises';
 import path from 'path';
 import url from 'url';
 import debug from 'debug';
-import picomatch from 'picomatch';
 import http from 'http';
 import express from 'express';
 import nunjucks from 'nunjucks';
@@ -44,7 +43,8 @@ import * as socketio from 'socket.io'
 
 import * as rl from './rl.js';
 import { data_proxy, status_proxy } from './rl.data.js';
-import { bytes_to_string, date_to_string, is_same_filesystem, system_poweroff, system_reboot } from './util.js';
+import { delete_data_file, get_data_path, get_data_file_info, get_log_path, validate_data_file } from './rl.files.js'
+import { is_same_filesystem, system_poweroff, system_reboot } from './util.js';
 
 const __filename = url.fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
@@ -155,30 +155,6 @@ app.get('/log/', async (request, reply) => {
 });
 
 
-// data and log file helper functions
-async function validate_data_file(filename) {
-    if (!is_valid_data_file(filename)) {
-        throw Error(`invalid data filename: ${filename}`);
-    }
-}
-
-function is_valid_data_file(filename) {
-    return filename.indexOf(path.sep) < 0;
-}
-
-async function delete_data_file(filename) {
-    return fs.unlink(get_data_path(filename));
-}
-
-function get_data_path(filename) {
-    return path.join(rl.path_data, filename);
-}
-
-function get_log_path() {
-    return rl.path_system_logfile;
-}
-
-
 // request handling helper functions
 function redirect_back_or_message(request, reply, message) {
     const referer = request.headers?.referer;
@@ -231,37 +207,6 @@ async function amend_system_context(context) {
         context.version_string = rl_version.version_string;
         context.version = rl_version.version;
     }
-}
-
-
-// data file info helper functions
-async function get_data_file_info() {
-    const files = await get_data_files();
-    const files_info = await Promise.all(files.map(get_file_info));
-    return sort_file_info(files_info);
-}
-
-async function get_data_files() {
-    const is_data_file = picomatch('*.@(rld|csv)');
-    const dir_files = await fs.readdir(rl.path_data);
-    const data_files = dir_files.filter(f => is_data_file(f));
-    return data_files.map(f => path.join(rl.path_data, f));
-}
-
-function sort_file_info(files_info) {
-    const compare_name = (a, b) => a.basename.localeCompare(b.basename);
-    return files_info.sort(compare_name);
-}
-
-async function get_file_info(filename) {
-    const file_stat = await fs.stat(filename);
-    const file_info = {
-        basename: path.basename(filename),
-        dirname: path.dirname(filename),
-        modified: date_to_string(file_stat.mtime),
-        size: bytes_to_string(file_stat.size),
-    };
-    return file_info;
 }
